fix(qa): extract acyclic runtime parity usage contract
