+++ extensions/qa-lab/src/runtime-parity-usage.ts
@@ -1,0 +1,7 @@
+export type RuntimeParityUsage = {
+  inputTokens: number;
+  outputTokens: number;
+  totalTokens: number;
+  cacheRead?: number;
+  cacheWrite?: number;
+};
